# chinese-laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Chinese Learning API",
        "description": "API endpoints for user authentication, translation, and user management",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://juansoft.online",
            "description": "Local API server"
        },
        {
            "url": "https://juansoft.online"
        }
    ],
    "paths": {
...
        "/api/translate/to-chinese": {
            "post": {
                "tags": [
                    "Translation"
                ],
                "summary": "Translate English text to Chinese",
                "operationId": "291bc80a10a72296a3c7a82be535e1bc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "text"
                                ],
                                "properties": {
                                    "text": {
                                        "type": "string",
                                        "example": "Hello"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful translation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "translation": {
                                            "type": "string",
                                            "example": "你好"
                                        },
                                        "pinyin": {
                                            "type": "string",
                                            "example": "ni hao"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/translate/to-english": {
            "post": {
                "tags": [
                    "Translation"
                ],
                "summary": "Translate Chinese text to English",
                "operationId": "692d4beecdb3ecc766767233ff6b7202",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "text"
                                ],
                                "properties": {
                                    "text": {
                                        "type": "string",
                                        "example": "你好"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful translation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "translation": {
                                            "type": "string",
                                            "example": "Hello"
                                        },
                                        "pinyin": {
                                            "type": "string",
                                            "example": "ni hao"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/pinyin": {
            "post": {
                "tags": [
                    "Translation"
                ],
                "summary": "Convert Chinese text to pinyin",
                "operationId": "9d4a8e75aa143f38b63925c43adf2e17",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "text"
                                ],
                                "properties": {
                                    "text": {
                                        "type": "string",
                                        "example": "你好"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful conversion",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "pinyin": {
                                            "type": "string",
                                            "example": "ni hao"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}